+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## action@0.0.4 - 2026-03-22
+
+- Remove custom `npm-token` parameter, favoring `.npmrc` or environment variables for NPM auth
+- Rename `release-url` output to `github-release-url` in Action
+- Improve `DomainError` and error formatting for better CLI/Action UX
+
 ## action@0.0.3 - 2026-03-22
 
 - Fix changelog matching rules